requirements.txt
Django>=4.2,<5.0
djangorestframework
celery
redis
django-redis
psycopg2-binary
ping3